icp completion events
